[Works] Player New Controller - Free Movement
Free movement for player (i.e. up, down, left, right), in a way that the player movement animation happens at the beat of the music. However, there is no restriction of player only moving if they press a key on beat.

Also started working on the Sandbox 4 for enemies; especifically the Radial Fireball Projectiles.